requirements.txt
pydantic>=2.6
python-dotenv>=1.0
orjson>=3.9
loguru>=0.7
regex>=2024.5
tqdm>=4.66
typer>=0.12
rdflib>=7.0
owlready2>=0.46
faiss-cpu>=1.8
sentence-transformers>=2.6
numpy>=1.26
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
tenacity>=8.3
pytest>=8.2
flashtext>=2.7
rapidfuzz>=3.9
neo4j>=5.21
